3 files representing Playlist.
A program that implements a playlist for songs. A Song class having Title and Duration for a song. The program have an Album class containing a list of songs. The albums is stored in an Arraylist. Songs from different albums can be added to the playlist and will appear in the list in the order they are added. Once the songs have been added to the playlist, there is a menu of options to:
        Quit;
       Skip forward to the next song;
       Skip backwards to a previous song;
       Replay the current song;
       List the songs in the playlist;
       Remove the current song.
       A song exist in an album before it can be added to the playlist (so you can only play songs that you own).